Django/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.3
attrs==20.3.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.2
django-admin-honeypot==1.1.0
django-allauth==0.44.0
django-braces==1.14.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-oauth-toolkit==1.5.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
idna==2.10
imbalanced-learn==0.8.0
imblearn==0.0
isodate==0.6.0
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
joblib==1.0.1
jwcrypto==0.8
lazy-object-proxy==1.6.0
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.2.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.4
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
smote==0.1
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.3
wrapt==1.12.1
zeep==4.0.0
